organize password-health.ts contents into new model files
